+++ server-spi-private/src/main/java/org/keycloak/events/admin/ResourceType.java
@@ -150,6 +150,11 @@ public enum ResourceType {
     /**
      *
      */
+    , CLIENT_SCOPE_POLICY
+
+    /**
+     *
+     */
     , CLIENT_SCOPE_MAPPING
 
     /**